refactor(frontend): extract shared SSE event handling and dedicated router
- Add useAgentEvents composable owning the agent SSE event state machine
  (message/tool/step/error/title/plan), previously duplicated (~170 lines)
  between ChatPage and SharePage
- Move routes and auth guard from main.ts to src/router with lazy-loaded
  pages so Monaco/NoVNC/Claw stay out of the initial bundle
- Drop the router import from api/client.ts to break the api <-> main
  circular dependency (401 redirect uses location.href only)
- Type chatWithSession attachments as {file_id, filename}[] to match the
  actual payload instead of string[]
- Remove dead code (ClawPage.handleRecreateClaw, unused imports) and fix
  setInterval/setTimeout typings surfaced by vue-tsc
- Pin typescript to 5.4.5: the previous ^5.1.3 range resolves to 5.8.x
  which crashes vue-tsc 1.8.27 on fresh installs
