+++ frontend/package.json
@@ -38,7 +38,7 @@
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.24",
     "tailwindcss": "^3.3.2",
-    "typescript": "^5.1.3",
+    "typescript": "5.4.5",
     "vite": "^4.3.9",
     "vite-plugin-monaco-editor": "^1.1.0",
     "vue-tsc": "^1.6.5"
