+++ frontend/src/pages/ChatPage.vue
@@ -129,22 +129,14 @@
 
 <script setup lang="ts">
 import SimpleBar from '../components/SimpleBar.vue';
-import { ref, onMounted, watch, nextTick, onUnmounted, reactive, toRefs } from 'vue';
+import { ref, onMounted, onUnmounted, reactive, toRefs } from 'vue';
 import { useRouter, onBeforeRouteUpdate } from 'vue-router';
 import { useI18n } from 'vue-i18n';
 import ChatBox from '../components/ChatBox.vue';
 import ChatMessage from '../components/ChatMessage.vue';
 import * as agentApi from '../api/agent';
-import { Message, MessageContent, ToolContent, StepContent, AttachmentsContent, isConsecutiveAssistant } from '../types/message';
-import {
-  StepEventData,
-  ToolEventData,
-  MessageEventData,
-  ErrorEventData,
-  TitleEventData,
-  PlanEventData,
-  AgentSSEEvent,
-} from '../types/event';
+import { MessageContent, AttachmentsContent, isConsecutiveAssistant } from '../types/message';
+import { AgentSSEEvent } from '../types/event';
 import ToolPanel from '../components/ToolPanel.vue'
 import PlanPanel from '../components/PlanPanel.vue';
 import { ArrowDown, FileSearch, PanelLeft, Lock, Globe, Link, Check } from 'lucide-vue-next';
@@ -154,6 +146,7 @@ import type { FileInfo } from '../api/file';
 import { useLeftPanel } from '../composables/useLeftPanel'
 import { useSessionFileList } from '../composables/useSessionFileList'
 import { useFilePanel } from '../composables/useFilePanel'
+import { useAgentEvents } from '../composables/useAgentEvents'
 import { copyToClipboard } from '../utils/dom'
 import { SessionStatus } from '../types/response';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
@@ -165,21 +158,36 @@ const { toggleLeftPanel, isLeftPanelShow } = useLeftPanel()
 const { showSessionFileList } = useSessionFileList()
 const { hideFilePanel } = useFilePanel()
 
-// Create initial state factory
+// Non-state refs that don't need reset
+const toolPanel = ref<InstanceType<typeof ToolPanel>>()
+const simpleBarRef = ref<InstanceType<typeof SimpleBar>>();
+const observerRef = ref<HTMLDivElement>();
+const chatContainerRef = ref<HTMLDivElement>();
+
+// Shared SSE event handling (messages, plan, title, tool tracking)
+const {
+  isLoading,
+  messages,
+  realTime,
+  follow,
+  title,
+  plan,
+  lastEventId,
+  handleEvent,
+  resetEvents,
+  handleToolClick,
+  jumpToRealTime,
+} = useAgentEvents({
+  liveTools: true,
+  showToolPanel: (tool, live) => toolPanel.value?.showToolPanel(tool, live),
+  scrollToBottom: () => simpleBarRef.value?.scrollToBottom(),
+});
+
+// Page-specific state (input, share settings, connection handle)
 const createInitialState = () => ({
   inputMessage: '',
-  isLoading: false,
   sessionId: undefined as string | undefined,
-  messages: [] as Message[],
   toolPanelSize: 0,
-  realTime: true,
-  follow: true,
-  title: t('New Chat'),
-  plan: undefined as PlanEventData | undefined,
-  lastNoMessageTool: undefined as ToolContent | undefined,
-  lastMessageTool: undefined as ToolContent | undefined,
-  lastTool: undefined as ToolContent | undefined,
-  lastEventId: undefined as string | undefined,
   cancelCurrentChat: null as (() => void) | null,
   attachments: [] as FileInfo[],
   shareMode: 'private' as 'private' | 'public', // Default to private mode
@@ -187,23 +195,12 @@ const createInitialState = () => ({
   sharingLoading: false // Loading state for share operations
 });
 
-// Create reactive state
 const state = reactive(createInitialState());
 
-// Destructure refs from reactive state
 const {
   inputMessage,
-  isLoading,
   sessionId,
-  messages,
   toolPanelSize,
-  realTime,
-  follow,
-  title,
-  plan,
-  lastNoMessageTool,
-  lastTool,
-  lastEventId,
   cancelCurrentChat,
   attachments,
   shareMode,
@@ -211,146 +208,16 @@ const {
   sharingLoading
 } = toRefs(state);
 
-// Non-state refs that don't need reset
-const toolPanel = ref<InstanceType<typeof ToolPanel>>()
-const simpleBarRef = ref<InstanceType<typeof SimpleBar>>();
-const observerRef = ref<HTMLDivElement>();
-const chatContainerRef = ref<HTMLDivElement>();
-
-// Reset all refs to their initial values
+// Reset all state to initial values
 const resetState = () => {
   // Cancel any existing chat connection
   if (cancelCurrentChat.value) {
     cancelCurrentChat.value();
   }
 
-  // Reset reactive state to initial values
   Object.assign(state, createInitialState());
+  resetEvents();
 };
-
-// Watch message changes and automatically scroll to bottom
-watch(messages, async () => {
-  await nextTick();
-  if (follow.value) {
-    simpleBarRef.value?.scrollToBottom();
-  }
-}, { deep: true });
-
-
-
-const getLastStep = (): StepContent | undefined => {
-  return messages.value.filter(message => message.type === 'step').pop()?.content as StepContent;
-}
-
-// Handle message event
-const handleMessageEvent = (messageData: MessageEventData) => {
-  messages.value.push({
-    type: messageData.role,
-    content: {
-      ...messageData
-    } as MessageContent,
-  });
-
-  if (messageData.attachments?.length > 0) {
-    messages.value.push({
-      type: 'attachments',
-      content: {
-        ...messageData
-      } as AttachmentsContent,
-    });
-  }
-}
-
-// Handle tool event
-const handleToolEvent = (toolData: ToolEventData) => {
-  const lastStep = getLastStep();
-  let toolContent: ToolContent = {
-    ...toolData
-  }
-  if (lastTool.value && lastTool.value.tool_call_id === toolContent.tool_call_id) {
-    Object.assign(lastTool.value, toolContent);
-  } else {
-    if (lastStep?.status === 'running') {
-      lastStep.tools.push(toolContent);
-    } else {
-      messages.value.push({
-        type: 'tool',
-        content: toolContent,
-      });
-    }
-    lastTool.value = toolContent;
-  }
-  if (toolContent.name !== 'message') {
-    lastNoMessageTool.value = toolContent;
-    if (realTime.value) {
-      toolPanel.value?.showToolPanel(toolContent, true);
-    }
-  }
-}
-
-// Handle step event
-const handleStepEvent = (stepData: StepEventData) => {
-  const lastStep = getLastStep();
-  if (stepData.status === 'running') {
-    messages.value.push({
-      type: 'step',
-      content: {
-        ...stepData,
-        tools: []
-      } as StepContent,
-    });
-  } else if (stepData.status === 'completed') {
-    if (lastStep) {
-      lastStep.status = stepData.status;
-    }
-  } else if (stepData.status === 'failed') {
-    isLoading.value = false;
-  }
-}
-
-// Handle error event
-const handleErrorEvent = (errorData: ErrorEventData) => {
-  isLoading.value = false;
-  messages.value.push({
-    type: 'assistant',
-    content: {
-      content: errorData.error,
-      timestamp: errorData.timestamp
-    } as MessageContent,
-  });
-}
-
-// Handle title event
-const handleTitleEvent = (titleData: TitleEventData) => {
-  title.value = titleData.title;
-}
-
-// Handle plan event
-const handlePlanEvent = (planData: PlanEventData) => {
-  plan.value = planData;
-}
-
-// Main event handler function
-const handleEvent = (event: AgentSSEEvent) => {
-  if (event.event === 'message') {
-    handleMessageEvent(event.data as MessageEventData);
-  } else if (event.event === 'tool') {
-    handleToolEvent(event.data as ToolEventData);
-  } else if (event.event === 'step') {
-    handleStepEvent(event.data as StepEventData);
-  } else if (event.event === 'done') {
-    //isLoading.value = false;
-  } else if (event.event === 'wait') {
-    // TODO: handle wait event
-  } else if (event.event === 'error') {
-    handleErrorEvent(event.data as ErrorEventData);
-  } else if (event.event === 'title') {
-    handleTitleEvent(event.data as TitleEventData);
-  } else if (event.event === 'plan') {
-    handlePlanEvent(event.data as PlanEventData);
-  }
-  lastEventId.value = event.data.event_id;
-}
 
 const handleSubmit = () => {
   chat(inputMessage.value, attachments.value);
@@ -498,37 +365,6 @@ onUnmounted(() => {
     cancelCurrentChat.value = null;
   }
 })
-
-const isLastNoMessageTool = (tool: ToolContent) => {
-  return tool.tool_call_id === lastNoMessageTool.value?.tool_call_id;
-}
-
-const isLiveTool = (tool: ToolContent) => {
-  if (tool.status === 'calling') {
-    return true;
-  }
-  if (!isLastNoMessageTool(tool)) {
-    return false;
-  }
-  if (tool.timestamp > Date.now() - 5 * 60 * 1000) {
-    return true;
-  }
-  return false;
-}
-
-const handleToolClick = (tool: ToolContent) => {
-  realTime.value = false;
-  if (sessionId.value) {
-    toolPanel.value?.showToolPanel(tool, isLiveTool(tool));
-  }
-}
-
-const jumpToRealTime = () => {
-  realTime.value = true;
-  if (lastNoMessageTool.value) {
-    toolPanel.value?.showToolPanel(lastNoMessageTool.value, isLiveTool(lastNoMessageTool.value));
-  }
-}
 
 const handleFollow = () => {
   follow.value = true;
