+++ frontend/src/pages/SharePage.vue
@@ -92,29 +92,20 @@
 
 <script setup lang="ts">
 import SimpleBar from '../components/SimpleBar.vue';
-import { ref, onMounted, onUnmounted, watch, nextTick, reactive, toRefs } from 'vue';
+import { ref, onMounted, onUnmounted, reactive, toRefs } from 'vue';
 import { useRouter } from 'vue-router';
 import { useI18n } from 'vue-i18n';
 import ChatMessage from '../components/ChatMessage.vue';
 import * as agentApi from '../api/agent';
-import { Message, MessageContent, ToolContent, StepContent, AttachmentsContent, isConsecutiveAssistant } from '../types/message';
-import {
-  StepEventData,
-  ToolEventData,
-  MessageEventData,
-  ErrorEventData,
-  TitleEventData,
-  PlanEventData,
-  AgentSSEEvent,
-} from '../types/event';
+import { isConsecutiveAssistant } from '../types/message';
 import ToolPanel from '../components/ToolPanel.vue'
 import PlanPanel from '../components/PlanPanel.vue';
 import { ArrowDown, FileSearch, Link, Bot } from 'lucide-vue-next';
 import ManusLogoTextIcon from '../components/icons/ManusLogoTextIcon.vue';
 import { showErrorToast, showSuccessToast } from '../utils/toast';
-import type { FileInfo } from '../api/file';
 import { useSessionFileList } from '../composables/useSessionFileList'
 import { useFilePanel } from '../composables/useFilePanel'
+import { useAgentEvents } from '../composables/useAgentEvents'
 import LoadingIndicator from '@/components/ui/LoadingIndicator.vue';
 import { copyToClipboard } from '../utils/dom'
 
@@ -123,183 +114,55 @@ const { t } = useI18n()
 const { showSessionFileList } = useSessionFileList()
 const { hideFilePanel } = useFilePanel()
 
-// Create initial state factory
+// Non-state refs that don't need reset
+const toolPanel = ref<InstanceType<typeof ToolPanel>>()
+const simpleBarRef = ref<InstanceType<typeof SimpleBar>>();
+let countdownTimer: number | null = null;
+
+// Shared SSE event handling (messages, plan, title, tool tracking).
+// Replays never show live tools.
+const {
+  isLoading,
+  messages,
+  realTime,
+  follow,
+  title,
+  plan,
+  handleEvent,
+  resetEvents,
+  handleToolClick,
+  jumpToRealTime,
+} = useAgentEvents({
+  liveTools: false,
+  showToolPanel: (tool, live) => toolPanel.value?.showToolPanel(tool, live),
+  scrollToBottom: () => simpleBarRef.value?.scrollToBottom(),
+});
+
+// Page-specific state (replay overlay and progress)
 const createInitialState = () => ({
-  inputMessage: '',
-  isLoading: false,
   sessionId: undefined as string | undefined,
-  messages: [] as Message[],
   toolPanelSize: 0,
-  realTime: true,
-  follow: true,
-  title: t('New Chat'),
-  plan: undefined as PlanEventData | undefined,
-  lastNoMessageTool: undefined as ToolContent | undefined,
-  lastMessageTool: undefined as ToolContent | undefined,
-  lastTool: undefined as ToolContent | undefined,
-  lastEventId: undefined as string | undefined,
-  attachments: [] as FileInfo[],
   showReplayOverlay: false,
   countdown: 3,
   jumpToEnd: false,
   replayCompleted: false,
 });
 
-// Create reactive state
 const state = reactive(createInitialState());
 
-// Destructure refs from reactive state
 const {
-  isLoading,
   sessionId,
-  messages,
   toolPanelSize,
-  realTime,
-  follow,
-  title,
-  plan,
-  lastNoMessageTool,
-  lastTool,
-  lastEventId,
   showReplayOverlay,
   countdown,
   jumpToEnd,
   replayCompleted,
 } = toRefs(state);
 
-// Non-state refs that don't need reset
-const toolPanel = ref<InstanceType<typeof ToolPanel>>()
-const simpleBarRef = ref<InstanceType<typeof SimpleBar>>();
-let countdownTimer: number | null = null;
-
-// Watch message changes and automatically scroll to bottom
-watch(messages, async () => {
-  await nextTick();
-  if (follow.value) {
-    simpleBarRef.value?.scrollToBottom();
-  }
-}, { deep: true });
-
-
-
-const getLastStep = (): StepContent | undefined => {
-  return messages.value.filter(message => message.type === 'step').pop()?.content as StepContent;
-}
-
-// Handle message event
-const handleMessageEvent = (messageData: MessageEventData) => {
-  messages.value.push({
-    type: messageData.role,
-    content: {
-      ...messageData
-    } as MessageContent,
-  });
-
-  if (messageData.attachments?.length > 0) {
-    messages.value.push({
-      type: 'attachments',
-      content: {
-        ...messageData
-      } as AttachmentsContent,
-    });
-  }
-}
-
-// Handle tool event
-const handleToolEvent = (toolData: ToolEventData) => {
-  const lastStep = getLastStep();
-  let toolContent: ToolContent = {
-    ...toolData
-  }
-  if (lastTool.value && lastTool.value.tool_call_id === toolContent.tool_call_id) {
-    Object.assign(lastTool.value, toolContent);
-  } else {
-    if (lastStep?.status === 'running') {
-      lastStep.tools.push(toolContent);
-    } else {
-      messages.value.push({
-        type: 'tool',
-        content: toolContent,
-      });
-    }
-    lastTool.value = toolContent;
-  }
-  if (toolContent.name !== 'message') {
-    lastNoMessageTool.value = toolContent;
-    if (realTime.value) {
-      toolPanel.value?.showToolPanel(toolContent, false);
-    }
-  }
-}
-
-// Handle step event
-const handleStepEvent = (stepData: StepEventData) => {
-  const lastStep = getLastStep();
-  if (stepData.status === 'running') {
-    messages.value.push({
-      type: 'step',
-      content: {
-        ...stepData,
-        tools: []
-      } as StepContent,
-    });
-  } else if (stepData.status === 'completed') {
-    if (lastStep) {
-      lastStep.status = stepData.status;
-    }
-  } else if (stepData.status === 'failed') {
-    isLoading.value = false;
-  }
-}
-
-// Handle error event
-const handleErrorEvent = (errorData: ErrorEventData) => {
-  isLoading.value = false;
-  messages.value.push({
-    type: 'assistant',
-    content: {
-      content: errorData.error,
-      timestamp: errorData.timestamp
-    } as MessageContent,
-  });
-}
-
-// Handle title event
-const handleTitleEvent = (titleData: TitleEventData) => {
-  title.value = titleData.title;
-}
-
-// Handle plan event
-const handlePlanEvent = (planData: PlanEventData) => {
-  plan.value = planData;
-}
-
-// Main event handler function
-const handleEvent = (event: AgentSSEEvent) => {
-  if (event.event === 'message') {
-    handleMessageEvent(event.data as MessageEventData);
-  } else if (event.event === 'tool') {
-    handleToolEvent(event.data as ToolEventData);
-  } else if (event.event === 'step') {
-    handleStepEvent(event.data as StepEventData);
-  } else if (event.event === 'done') {
-    //isLoading.value = false;
-  } else if (event.event === 'wait') {
-    // TODO: handle wait event
-  } else if (event.event === 'error') {
-    handleErrorEvent(event.data as ErrorEventData);
-  } else if (event.event === 'title') {
-    handleTitleEvent(event.data as TitleEventData);
-  } else if (event.event === 'plan') {
-    handlePlanEvent(event.data as PlanEventData);
-  }
-  lastEventId.value = event.data.event_id;
-}
-
-// Reset all refs to their initial values
+// Reset all state to initial values
 const resetState = () => {
-  // Reset reactive state to initial values
   Object.assign(state, createInitialState());
+  resetEvents();
 };
 
 const replay = async () => {
@@ -385,20 +248,6 @@ onUnmounted(() => {
     countdownTimer = null;
   }
 });
-
-const handleToolClick = (tool: ToolContent) => {
-  realTime.value = false;
-  if (sessionId.value) {
-    toolPanel.value?.showToolPanel(tool, false);
-  }
-}
-
-const jumpToRealTime = () => {
-  realTime.value = true;
-  if (lastNoMessageTool.value) {
-    toolPanel.value?.showToolPanel(lastNoMessageTool.value, false);
-  }
-}
 
 const handleFollow = () => {
   follow.value = true;
